@@ -1,4 +1,4 @@
-﻿cmake_minimum_required (VERSION 3.8)
+﻿cmake_minimum_required (VERSION 3.15)
 
 project (ecs CXX)
 
@@ -10,16 +10,23 @@ install(
     DESTINATION include
 )
 
-# Require c++17
-set(CMAKE_CXX_STANDARD 17)
+# Require c++20 and concepts
+set(CMAKE_CXX_STANDARD 20)
 set(CXX_STANDARD_REQUIRED on)
-
 
 # determine whether this is a standalone project or included by other projects
 set(ECS_STANDALONE_PROJECT OFF)
 if (CMAKE_CURRENT_SOURCE_DIR STREQUAL CMAKE_SOURCE_DIR)
     set(ECS_STANDALONE_PROJECT ON)
 endif ()
+
+if (MSVC)
+    # warning level 4, modules
+    target_compile_options(ecs INTERFACE /W4)
+else()
+    # lots of warnings
+    target_compile_options(ecs INTERFACE -Wall -Wextra -pedantic -fconcepts)
+endif()
 
 # Project headers
 # add include folders to the library and targets that consume it
@@ -39,9 +46,6 @@ else()
         >
     )
 endif()
-
-# Include GSL
-add_subdirectory ("GSL")
 
 if (ECS_STANDALONE_PROJECT)
     # Benchmark
